backend/requirement.txt
fastapi
uvicorn
python-multipart
rembg
pillow
numpy
pydantic-settings
onnxruntime